Support modtime before I forget
